server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-CORS
Flask-Limiter
Flask-Mail
psycopg2-binary
redis
Authlib
requests
PyJWT
python-dotenv
gunicorn
python-dateutil
pyotp
flask-jwt-extended
qrcode
pillow
bcrypt
python-multipart
flask-limiter
marshmallow